feat: conditional patch insert class methods
