+++ instrumentation/active_record/lib/opentelemetry/instrumentation/active_record/patches/persistence_insert_class_methods.rb
@@ -1,0 +1,59 @@
+# frozen_string_literal: true
+
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+module OpenTelemetry
+  module Instrumentation
+    module ActiveRecord
+      module Patches
+        # Module to prepend to ActiveRecord::Persistence::ClassMethods for instrumentating
+        # insert/upsert class methods added in Rails 6.0
+        module PersistenceInsertClassMethods
+          def insert(attributes, returning: nil, unique_by: nil)
+            tracer.in_span("#{self}.insert") do
+              super
+            end
+          end
+
+          def insert_all(attributes, returning: nil, unique_by: nil)
+            tracer.in_span("#{self}.insert_all") do
+              super
+            end
+          end
+
+          def insert!(attributes, returning: nil)
+            tracer.in_span("#{self}.insert!") do
+              super
+            end
+          end
+
+          def insert_all!(attributes, returning: nil)
+            tracer.in_span("#{self}.insert_all!") do
+              super
+            end
+          end
+
+          def upsert(attributes, returning: nil, unique_by: nil)
+            tracer.in_span("#{self}.upsert") do
+              super
+            end
+          end
+
+          def upsert_all(attributes, returning: nil, unique_by: nil)
+            tracer.in_span("#{self}.upsert_all") do
+              super
+            end
+          end
+
+          private
+
+          def tracer
+            ActiveRecord::Instrumentation.instance.tracer
+          end
+        end
+      end
+    end
+  end
+end
